+++ Packages/Vanilla Normal/Samples~/Normal Sample/NormalSample.cs
@@ -1,6 +1,3 @@
-using System.Collections;
-using System.Collections.Generic;
-
 using UnityEngine;
 
 using Vanilla.TypeMenu;
@@ -19,14 +16,7 @@ namespace Vanilla
         public Normal normal;
 
         [SerializeField]
-        public UnsafeNormal unsafeNormal;
-
-        [SerializeField]
         public EasingNormal easingNormal;
-
-        [SerializeField]
-        public UnsafeEasingNormal unsafeEasingNormal;
-
 
         void Start()
         {
@@ -38,19 +28,13 @@ namespace Vanilla
                 swappableNormal.Fill(conditional: null);
             }
 
-            normal.OnChange             += n => Debug.Log(message: $"The regular normal changed to [{n}]");
-            unsafeNormal.OnChange       += n => Debug.Log(message: $"The unsafe normal changed to [{n}]");
-            easingNormal.OnChange       += n => Debug.Log(message: $"The easing normal changed to [{n}]");
-            unsafeEasingNormal.OnChange += n => Debug.Log(message: $"The unsafe easing normal changed to [{n}]");
+            normal.OnChange             += n => Debug.Log(message: $"The normal changed to [{n}]");
+            easingNormal.OnChange       += n => Debug.Log(message: $"The eased normal changed to [{n}]");
 
             normal.Full.onTrue += () => Debug.Log(message: "The normal is full!");
 
-            unsafeNormal.Full.onFalse += () => Debug.Log(message: "The unsafe normal is no longer full!");
-
             easingNormal.OnIncrease += n => Debug.Log(message: $"The easing Normal normal increased to [{n}]!");
-
-            unsafeEasingNormal.Empty.onChange += isEmpty => Debug.Log(message: $"The unsafe easing normal is {(isEmpty ? string.Empty : "not ")}empty");
-
+            
         }
 
     }