+++ Packages/Vanilla Normal/CHANGELOG.md
@@ -1,9 +1,18 @@
-# Vanilla Package Changelog
+# Vanilla Normal Changelog
 
 All notable changes to this package will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
+
+## [1.1.0] - 09-03-2022
+
+### Changed
+- Support for Danger is now built into Normal and EasingNormal automatically. They will opt-in to use BitwiseEquals if Danger is detected.
+
+### Removed
+- UnsafeNormal has been deprecated.
+- UnsafeEasingNormal has been deprecated.
 
 ## [1.0.0] - 06-06-2021